\name{labtrecasparnitin-package}
\alias{labtrecasparnitin-package}
\alias{labtrecasparnitin}
\docType{package}
\title{
\packageTitle{labtrecasparnitin}
}
\description{
\packageDescription{labtrecasparnitin}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{labtrecasparnitin}
\packageIndices{labtrecasparnitin}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{labtrecasparnitin}

Maintainer: \packageMaintainer{labtrecasparnitin}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
